unreal: add support to transcode ktx2 textures
Add KTX2 transcoding support for supported UE compressed formats.
This allows for UE to harness the full benefits of runtime KTX2
compressed formats that will yield smaller texture footprints at
minimal runtime transcoding performance cost.

The existing flag ForcePixelFormat will be used to determine
the target format to transcode to. If the flag is not set, we will skip
the KTX2 MIPs transcoding and default to gltfruntime behavior.

A table of supported formats is defined in GetMappingTable. This list
is non-exhaustive and can be extended with more testing and validation
capacity.